requirements.txt
agora-token-builder==1.0.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
boto3==1.34.151
botocore==1.34.151
cachetools==5.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.0.3
colorama==0.4.4
comtypes==1.1.11
distlib==0.3.8
Django==4.2.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
emoji==2.12.1
filelock==3.13.1
Flask==2.0.2
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==1.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
pillow==10.3.0
platformdirs==4.2.0
playsound==1.3.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
PyAudio==0.2.12
PyJWT==2.8.0
PyPDF2==2.6.0
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.90
pywin32==304
razorpay==1.4.2
requests==2.28.1
rsa==4.9
s3transfer==0.10.2
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
typing-extensions==4.12.1
tzdata==2022.1
urllib3==1.26.10
virtualenv==20.25.0
Werkzeug==2.0.2
wikipedia==1.4.0
